@@ -3,15 +3,17 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "geohashing": "^1.1.3",
     "random-js": "^2.1.0"
   },
   "engines": {
-    "node": ">=12.0.0 <=18.0.0"
+    "node": ">=18.0.0 <=20.0.0"
   }
 }
