@@ -1,4 +1,5 @@
-const fs = require("fs");
+import { writeFileSync } from "fs";
+import { encodeBase32, decodeBase32 } from "geohashing";
 
 function gen(minLong = 1, minLat = 1, maxLong = 1.2, maxLat = 1.2) {
   const long = Number(
@@ -9,35 +10,57 @@ function gen(minLong = 1, minLat = 1, maxLong = 1.2, maxLat = 1.2) {
   return [long, lat];
 }
 
-const bbox = [34.2654333839, 29.5013261988, 35.8363969256, 33.2774264593];
-const cordsArr = new Map();
+function generateGeoData() {
+  const bbox = [34.2654333839, 29.5013261988, 35.8363969256, 33.2774264593];
+  const cordsArr = new Map();
 
-for (let i = 0; i < 1000; i++) {
-  let res;
+  for (let i = 0; i < 1000; i++) {
+    let res;
 
-  do {
-    res = gen(bbox[0], bbox[1], bbox[2], bbox[3]);
-  } while (cordsArr.has(res));
+    do {
+      res = gen(bbox[0], bbox[1], bbox[2], bbox[3]);
+    } while (cordsArr.has(res));
 
-  cordsArr.set(res, 1);
-}
+    cordsArr.set(res, 1);
+  }
 
-const geoData = {
-  type: "FeatureCollection",
-  features: [],
-};
-
-for (const cords of Array.from(cordsArr.keys())) {
-  const feature = {
-    type: "Feature",
-    properties: {},
-    geometry: {
-      type: "Point",
-      coordinates: cords,
-    },
+  const geoData = {
+    type: "FeatureCollection",
+    features: [],
   };
 
-  geoData.features.push(feature);
+  for (const cords of Array.from(cordsArr.keys())) {
+    const feature = {
+      type: "Feature",
+      properties: {},
+      geometry: {
+        type: "Point",
+        coordinates: cords,
+      },
+    };
+
+    geoData.features.push(feature);
+  }
+
+  writeFileSync("geoData.json", JSON.stringify(geoData, null, 2));
 }
 
-fs.writeFileSync("geoData.json", JSON.stringify(geoData, null, 2));
+function generateGeoHashJsonOut(count = 0) {
+  const bbox = [34.2654333839, 29.5013261988, 35.8363969256, 33.2774264593];
+  const cordsArr = [];
+
+  for (let i = 0; i < count; i++) {
+    let res;
+
+    do {
+      res = gen(bbox[0], bbox[1], bbox[2], bbox[3]);
+    } while (cordsArr.includes(res));
+
+    cordsArr.push(encodeBase32(res[1], res[0]));
+  }
+
+  console.log(JSON.stringify(cordsArr));
+}
+
+// generateGeoData();
+generateGeoHashJsonOut(1000);
